train sweep + train and test one model
