+++ .gitignore
@@ -79,6 +79,7 @@ target/
 # exclude data from source control by default
 /data/
 /models/
+/wandb/
 
 # Mac OS-specific storage files
 .DS_Store
