+++ conf/conf_train.yaml
@@ -1,0 +1,5 @@
+optimizer: 'sgd'
+lr: 0.00001
+batch_size:  256
+epochs: 10
+dropout: 0.3
