+++ src/models/train_utils.py
@@ -1,10 +1,4 @@
-import random
-import wandb
-import numpy as np
-import pandas as pd
 from tqdm import tqdm
-import matplotlib.pyplot as plt
-import yaml
 
 import torch
 import torch.nn as nn
@@ -16,6 +10,7 @@ from src.models.model import BertSentiment
 
 # set device to Apple M1 GPU if available
 device = "mps" if torch.backends.mps.is_available() else "cpu"
+loss_fn = nn.CrossEntropyLoss().to(device)
 
 def get_dataloaders(train_set, val_set, bs):
   trainloader = DataLoader(train_set, batch_size=bs, shuffle=True, num_workers=0)
@@ -32,7 +27,7 @@ def get_optimizer(model, lr, optimizer = 'adam'):
   elif optimizer == 'sgd':
     return optim.SGD(model.parameters(), momentum=0.9, lr=lr)  
 
-def train_one_epoch(model, trainloader, optimizer, loss_fn):
+def train_one_epoch(model, trainloader, optimizer):
   
   train_loss = 0
   n_correct = 0
@@ -65,7 +60,7 @@ def train_one_epoch(model, trainloader, optimizer, loss_fn):
   
   return loss_epoch, acc
 
-def evaluate_one_epoch(model, valloader, loss_fn):
+def evaluate_one_epoch(model, valloader):
   
   test_loss = 0
   n_correct = 0
@@ -92,64 +87,3 @@ def evaluate_one_epoch(model, valloader, loss_fn):
   acc = (n_correct/total)*100
   
   return loss_epoch, acc
-
-
-# load hyper parameters to sweep over from config file
-with open('src/models/conf/conf_train.yaml') as file:
-      config = yaml.load(file, Loader=yaml.FullLoader)
-
-# initialize wandb sweep
-sweep_id = wandb.sweep(sweep=config, project='mlops_dtu')
-
-def train():
-  
-  # initialize wndb
-  run = wandb.init(config=config)
-  
-  # define hyper parameters for wandb sweep
-  lr  =  wandb.config.lr
-  dropout = wandb.config.dropout
-  optimizer = wandb.config.optimizer
-  
-  # define constant hyper parametners
-  bs = 256
-  epochs = 5
-  
-  # read data files from path
-  train_set = torch.load('data/processed/train_set.pth')
-  val_set = torch.load('data/processed/val_set.pth')
-  
-  # create data loaders
-  trainloader, valloader = get_dataloaders(train_set, val_set, bs = bs)
-
-  # init model
-  model = get_model(dropout=dropout)
-  
-  # define criterion and optimizer
-  loss_fn = nn.CrossEntropyLoss().to(device)
-  optimizer = get_optimizer(model, lr, optimizer=optimizer)
-  
-  for e in range(epochs):
-    print(f'[EPOCH]: {e+1:3d}')
-
-    train_loss, train_acc = train_one_epoch(model, trainloader, optimizer, loss_fn)
-    val_loss, val_acc = evaluate_one_epoch(model, valloader, loss_fn)
-    
-    wandb.log({
-        'epoch': e, 
-        'train_acc': train_acc,
-        'train_loss': train_loss, 
-        'val_acc': val_acc, 
-        'val_loss': val_loss
-      })
-
-# Start sweep
-wandb.agent(sweep_id, function=train, count=10)
-  
-
-  
-  
-  
-  
-  
-
