+++ wandb/latest-run
@@ -1,1 +1,1 @@
-run-20221105_184649-u8tduiof
+run-20221106_095241-1omkeal1
